% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cauchy}
\alias{cauchy}
\title{Cauchy model}
\usage{
cauchy(alpha, beta)
}
\arguments{
\item{alpha}{The alpha parameter proportional to 
the fractal dimension of the time series.}

\item{beta}{The beta parameter proportional to the 
Hurst parameeter of the time series.}
}
\value{
A Cauchy process model.
}
\description{
This is a wrapper for the Cauchy process model 
in the RandomFields package. The parameters 
alpha and beta are smoothness and power scaling 
parameters. The time series has autocovariance 
function C(h) = (1 + r^alpha)^(-beta/alpha).
Fractal dimension D = alpha + 1 - alpha/2, alpha in (0,2]
Hurst parameter : H = 1 - beta/2, beta > 0.
}

