% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{farima}
\alias{farima}
\title{Create a FARIMA model.}
\usage{
farima(ar, ma, d)
}
\arguments{
\item{ar}{Autoregressive parameters.}

\item{ma}{Moving average parameters.}

\item{d}{Long term memory parameter.}
}
\value{
A parametrized FARIMA model.
}
\description{
Create a FARIMA model.
}

