% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fBm}
\alias{fBm}
\title{Create a fractional Brownian motion function model.}
\usage{
fBm(H)
}
\arguments{
\item{H}{The Hurst exponent}
}
\value{
A model for fractional Brownian motion.
}
\description{
Create a fractional Brownian motion function model.
}

