% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen}
\alias{gen}
\title{Generate a function from a model}
\usage{
gen(mod)
}
\arguments{
\item{mod}{A function model}
}
\value{
Returns a function based on the model 
 parameters in mod
}
\description{
Generate a function from a model
}

