% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-jitter.R
\name{jitter_params}
\alias{jitter_params}
\title{Jitter the parameters of a time series model}
\usage{
jitter_params(mod, percent)
}
\arguments{
\item{mod}{A time series model.}

\item{percent}{Max percentage change in parameter values.}
}
\value{
A times series with randonly perturbed. 
        parameters in mod
}
\description{
Jitter the parameters of a time series model
}

