% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{mackeyglass}
\alias{mackeyglass}
\title{Create a Mackey-Glass model}
\usage{
mackeyglass(tau = 17, beta = 0.25, gamma = 0.1, n = 10, init = 0.5,
  noise = 0)
}
\arguments{
\item{tau}{The time lag}

\item{beta}{The beta parameter}

\item{gamma}{The gamma parameter}

\item{n}{The n parameter}

\item{init}{A value used to initialize the discrete Mackeyglass
equation}

\item{noise}{The noise ratio to add to the function}
}
\value{
A model for a Mackey-Glass equation
}
\description{
Create a Mackey-Glass model
}

