% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{quadratic}
\alias{quadratic}
\title{Create a quadratic function model.}
\usage{
quadratic(x0)
}
\arguments{
\item{x0}{Initial value (TODO: ).}
}
\value{
A quadratic function model.
}
\description{
Create a quadratic function model.
}

