% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-jitter.R
\name{rsign}
\alias{rsign}
\title{Create random sample from {-1, 1}}
\usage{
rsign(n)
}
\arguments{
\item{n}{The number of samples.}
}
\value{
An n-length vector from {-1, 1}.
}
\description{
Uses stats::rbinom to create a random {-1, 1} vector.
}

