% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-suite.R
\name{sim_names}
\alias{sim_names}
\title{Returns the names corresonding to functions of \code{sim_suite()}}
\usage{
sim_names()
}
\value{
A vector of function names
}
\description{
Returns the names corresonding to functions of \code{sim_suite()}
}

