% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-suite.R
\name{sim_suite}
\alias{sim_suite}
\title{Generate A Default suite of function models.}
\usage{
sim_suite()
}
\value{
A list of two lists of function models with
         the same functions and varying default parameters.
}
\description{
Generate A Default suite of function models.
}

