% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{weierstrass}
\alias{weierstrass}
\title{Single parameter Weierstrass function model.}
\usage{
weierstrass(a, b = 0, c = 2, random = TRUE, density = 1)
}
\arguments{
\item{a}{The holder constant of the function. Equal to 
-log(a)/log(b) for the two parameter Weierstrass.
If b is zero, this is the amplitude pameter of the
two parameter Weierstrass function.}

\item{b}{The frequencies parameter. 
If zero the single parameter Weierstrass function is used.}

\item{c}{The amplitude parameter.}

\item{random}{If true generates a random phase model.}

\item{density}{Density of the grid on which the function is computed.
For example, if density = 100 then on (0,1) 100 points
are computed.}
}
\value{
The parametrized model.
}
\description{
Creates a model for a Weierstrass or random 
 phase Weierstrass function with single parameter
 a. The resulting function is a-Holder continuous.
 If random, theu function is evaluated on (1,n), 
 otherwise on (0,1),
}

